--continue option

+++ scripts/linux-build.sh
@@ -9,14 +9,16 @@ fi
 
 RUN_PKGS=''
 FORCE=''
+SKIP_BOOTSTRAP=1
 
 while (( "$#" )); do
     if [ "x$1" = x--tools ]; then RUN_TOOLS=1;
     elif [ "x$1" = x--all ]; then RUN_ALL=1;
     elif [ "x$1" = x--force ]; then FORCE='-f';
+    elif [ "x$1" = x--continue ]; then SKIP_BOOTSTRAP='';
     elif [ "x$1" = x-h -o "x$1" = x--help ]; then
         echo ''
-        echo " Usage: $0 [-h|--help] --force [--tools | --all] <pkgs. list>"
+        echo " Usage: $0 [-h|--help] --force --continue [--tools | --all] <pkgs. list>"
         echo ''
         echo " Default is --base (r-base-dev), tools include emacs and subversion."
         echo ''
@@ -71,9 +73,11 @@ echo "       FORCE: $FORCE"
 
 ## freetype and harfbuzz have a circular dependency
 ## and need to be bootstrapped in the order FT -> HB -> FT
-./build.sh $FORCE -p freetype
-./build.sh $FORCE -p harfbuzz
-rm -rf build/freetype-2.*
+if [ -z "$SKIP_BOOTSTRAP" ]; then
+    ./build.sh $FORCE -p freetype
+    ./build.sh $FORCE -p harfbuzz
+    rm -rf build/freetype-2.*
+fi
 
 ## required
 ./build.sh $FORCE -p r-base-dev
@@ -96,7 +100,7 @@ fi
 ## ./build.sh -f -p -- -k all
 
 echo "COPY stub configs"
-cp -r $PWD/stubs/pkgconfig-$OSHOST /usr/local/lib/pkgconfig-stubs
+cp -r $PWD/stubs/pkgconfig-$OSHOST /$PREFIX/lib/pkgconfig-stubs
 
 echo ''
 echo "=== DONE"
